+++ src/bsp_flat_scene.h
@@ -10,7 +10,7 @@ int bsp_flat_build_scene(BspFlatDraw *out, uint32_t capacity,
                          uint32_t vertex_srd_table_address,
                          float aspect_ratio);
 int bsp_flat_build_clear(BspFlatDraw *out, BspBundleVertex vertices[3],
-                         uint32_t indices[3],
+                         uint16_t indices[3],
                          uint32_t vertex_srd_table_address);
 int bsp_flat_update_camera(BspFlatDraw *draws, uint32_t draw_count,
                            const float position[3], const float forward[3],
